requirements.txt
python-stacktaskclient
python-keystoneclient
